Create config_loader to avoid circular imports
Update run_suite2p for calling config variables from config_loader
